requirements.txt
asyncio==3.4.3
dj-database-url==0.5.0
Django==3.0.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==9.1.0
django-heroku==0.3.1
gunicorn==20.0.4
html5lib==1.0.1
SQLAlchemy==1.2.17
whitenoise==5.0.1

psycopg2==2.8.4

djangorestframework==3.11.0
django-grappelli==2.14.2
django-filebrowser==3.13.1
django_summernote==0.8.11.6